+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/client/DefaultOAuth2TokenService.java
@@ -8,7 +8,8 @@ package com.sap.cloud.security.xsuaa.client;
 
 import static com.sap.cloud.security.xsuaa.client.OAuth2TokenServiceConstants.*;
 
-import com.sap.cloud.security.client.ApacheHttpClient4Adapter;
+import com.sap.cloud.security.client.ApacheHttpClient4Executor;
+import com.sap.cloud.security.client.CustomHttpClientAdapter;
 import com.sap.cloud.security.client.DefaultTokenClientConfiguration;
 import com.sap.cloud.security.client.SecurityHttpClient;
 import com.sap.cloud.security.client.SecurityHttpRequest;
@@ -59,7 +60,7 @@ public class DefaultOAuth2TokenService extends AbstractOAuth2TokenService {
    */
   @Deprecated(since = "4.0.0", forRemoval = true)
   public DefaultOAuth2TokenService(@Nonnull final CloseableHttpClient httpClient) {
-    this(new ApacheHttpClient4Adapter(httpClient));
+    this(new CustomHttpClientAdapter(new ApacheHttpClient4Executor(httpClient), httpClient::close));
   }
 
   /**
@@ -72,7 +73,7 @@ public class DefaultOAuth2TokenService extends AbstractOAuth2TokenService {
   public DefaultOAuth2TokenService(
       @Nonnull final CloseableHttpClient httpClient,
       @Nonnull final TokenCacheConfiguration tokenCacheConfiguration) {
-    this(new ApacheHttpClient4Adapter(httpClient), tokenCacheConfiguration);
+    this(new CustomHttpClientAdapter(new ApacheHttpClient4Executor(httpClient), httpClient::close), tokenCacheConfiguration);
   }
 
   @Override